fix(build): restore Debug/Release test parity; silence four warning classes
windows-x64-release never set AQUA_BUILD_C_API, so aqua_capi_test - the only
test covering the C API/JNI positional contract - was built and run in Debug
alone (443 vs 436 cases). The Android presets compile aqua_capi and
aqua_jni.cpp but run no tests at all, so that contract had a single
configuration behind it. With the switch on both configs report 443/443.

Note that cmake --build reuses the cache and does NOT pick up a preset
change; re-running cmake --preset <name> is required.

Warnings:
- tests: C4244 in jitter_control_replay (double -> uint32 ceiling),
  C4834 where EXPECT_THROW discarded a [[nodiscard]] result, C4245 for
  set_channels(-1)/set_sample_rate(-1) (now written as 0xFFFFFFFFu, the
  same value with the intent spelled out).
- clang-only, i.e. visible only when the JB debug-log macros are off:
  [[maybe_unused]] on the reanchor throttle field and on
  kJbRejectLogIntervalNs; #if around the startup_anchored/rejects lambda
  captures. The header field cannot use #if - the macros are PRIVATE to
  the core targets, so guarding them would give different class layouts
  across TUs (the offset of a private tail field would move).
- aqua_proto gets -Wno-nullability-extension under Clang: the 18 warnings
  come from protobuf/abseil headers and only appear without NDEBUG
  (Debug 18, Release 0); measured 18 -> 0.
- C4324 from audio_frame_queue's intentional alignas(64) (head_/tail_
  false-sharing split) is now suppressed locally under _MSC_VER with a
  comment, instead of being reported once per TU.

.gitignore: re-applied the .workbuddy/ / Testing/ / .kotlin/ / *.apk /
modern_cpp_roadmap.md entries - they were lost between the previous edit
and commit a50fead (26 files committed where 27 were expected).

Verified: Debug and Release build with 0 errors and pass 443/443; the four
warning codes are gone from a build that recompiled 27 TUs; the seven
affected translation units compile warning-free with the Android toolchain
in both configs.

+++ aqua_core/tests/audio/jitter_control_replay_test.cpp
@@ -710,7 +710,8 @@ TEST(JitterControlReplayTest, OutageResidueIsCappedNotLocked)
 
     const double stall_cap = aqua::config::JB_ADAPTIVE_STALL_PEAK_CAP_SLOTS;
     const double tail = mean_target_in(art, kOutage.run_ms - 2000.0, kOutage.run_ms - 100.0);
-    const auto settled = settled_at_ms(art, kOutage.outage_end_ms, stall_cap, 1000.0);
+    const auto settled = settled_at_ms(
+        art, kOutage.outage_end_ms, static_cast<std::uint32_t>(stall_cap), 1000.0);
     const auto& last = art.targets.back();
     // 按衰减律估算"峰值回落到封顶以下"所需时间（仅用于报告，不做断言）。
     const double release_estimate_s
